refactor: post component not properly setup
i was not fetching the post from the backend and was utilizing the post passed in from another parent. This caused an issue for the post detail view. It worked great at first with just loading the posts in the root, but i struggled getting render to pass in the post, there were just all kinds of bugs it seemed. Frustrated, i started again, this time using the backend to fetch the post, and simply passing in the id of the post i wanted or if not id was passed in, getting the id from match. This solved my issue, i just made sure functions such as vote caused me to update the posts state, still getting my head around, i do hope im doing this right :)